20191204_1522 meigarasにvolを追加、json変換に成功

+++ database/migrations/2019_10_14_211402_create_meigaras_table.php
@@ -24,7 +24,7 @@ class CreateMeigarasTable extends Migration
           $table->smallInteger('closingPrice')->nullable();
           $table->smallInteger('highPrice')->nullable();
           $table->smallInteger('lowPrice')->nullable();
-          $table->mediumInteger('volume')->nullable();
+          $table->integer('volume')->nullable();
           $table->integer('tradingValue')->nullable();
           $table->timestamps();
